# public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quotd API",
    "description": "Quotd REST API for managing AI voice interviews and case study generation.",
    "version": "1.0.0",
    "contact": {
      "name": "Quotd",
      "url": "https://quotd.io"
    }
  },
  "servers": [
    {
      "url": "https://app.quotd.io/api/v1",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key (qtd_...) as Bearer token"
      }
    },
    "schemas": {
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "url": { "type": "string", "format": "uri" },
          "events": { "type": "array", "items": { "type": "string" } },
          "secret": { "type": "string", "nullable": true },
          "active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "per_page": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/teams/{teamId}/members": {
      "get": {
        "summary": "List team members",
        "parameters": [{ "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Team members" } }
      },
      "post": {
        "summary": "Invite a team member",
        "parameters": [{ "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email", "role"], "properties": { "email": { "type": "string" }, "role": { "type": "string", "enum": ["editor", "viewer"] } } } } } },
        "responses": { "201": { "description": "Member added" } }
      },
      "delete": {
        "summary": "Remove a team member",
        "parameters": [{ "name": "teamId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["userId"], "properties": { "userId": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Member removed" } }
      }
    },
...
    "/profile/branding": {
      "get": {
        "summary": "Get branding settings",
        "responses": { "200": { "description": "Branding settings" } }
      },
      "patch": {
        "summary": "Update branding settings",
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "primary_color": { "type": "string" }, "welcome_message": { "type": "string", "nullable": true }, "logo_url": { "type": "string", "nullable": true } } } } } },
        "responses": { "200": { "description": "Branding updated" } }
      }
    },
...
    "/webhooks": {
      "get": {
        "summary": "List webhooks",
        "responses": { "200": { "description": "Webhooks", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Webhook" } }, "pagination": { "$ref": "#/components/schemas/Pagination" } } } } } } }
      },
      "post": {
        "summary": "Create webhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "events"],
                "properties": {
                  "url": { "type": "string", "format": "uri" },
                  "events": { "type": "array", "items": { "type": "string", "enum": ["interview.created", "interview.completed", "review.started", "review.completed", "draft.generated", "format.generated", "reminder.sent"] } },
                  "secret": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Webhook created" } }
      }
    },
...
  }
}